services/rag-pipeline/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
chromadb==0.4.18

# Git operations
GitPython==3.1.40

# ML/AI
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
numpy==1.24.3

# Code parsing
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.3
tree-sitter-typescript==0.20.4

# HTTP client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
